[WIP] Respect EditorConfig settings
This is a simple way to address https://github.com/prettier/prettier/issues/42.
It adds support for .editorconfig's `indent_style`, `tab_width`,
and `max_line_length` properties.

I still need to add caching and a proper async implementation,
but I think it's functionally correct.
It doesn't support the `end_of_line` property as described in
https://github.com/prettier/prettier/issues/42#issuecomment-327073326,
but that could be added later.