docs(ignore): mention ignore file in Ignoring Code
While working on #3321, I needed to ignore an entire file, but the "Ignoring Code" section of the docs only describes how to ignore blocks of code, so I wasn't aware of the default ignore file.

(The `.prettierignore` file is mentioned in the CLI section, but easily overlooked there.)

This commit extends the introduction sentence of the "Ignoring Code" section and adds a new subsection to explain how to ignore files.

https://github.com/prettier/prettier/issues/3321#issuecomment-350072566