Added prettier API for ignored files
The actual logic for creating an ignorer has been moved to `common/ignore.js`
so it can be shared easily between CLI and API.

The CLI has been updated to use the new `common/ignore.js` file by creating
the ignored via the CLI context.

The API has been added as a factory pattern so that reading the ignored file
contens is not repeated for each but only once at the creation of the
ignorer.