Added support to specify more than one ignore-path entry
Previously, passing multiple ignore paths was not something possible. So
if one wanted to use its .gitignore file as well as another file defining
extra ignored files that should however not be part of the .gitignore file,
then it was not possible to do so.

Now, the --ignore-path CLI option accepts being specified multiple times
to use multiple ignore paths.