fix(angular): Format <pre> in template strings properly.
The current logic doesn't take into account <pre> in a template string. This updates the logic to add additional base indentation to compare the alignment of the opening and closing pre tags if it's in a template string since it should be indented twice (once for the indentation of the template property itself, and once for the indentation of the string contents.

Because of the way <pre> works, this will only work properly if the closing </pre> tag was already indented to where it was supposed to be. Otherwise, the closing caret is dropped to the newline to align with the opening tag in order to keep whitespace consistent.