Fix formatting of lists in CSS property values
Fixes the following bug:

**Input:**
```
$my-list2:
  a // a
  b
  c;
```

**Actual Output:**
```
$my-list2: a// a
  bc;
```
(no space before comment, values after the comment glued together)

**Expected Output:**
```
$my-list2: a // a
  b c;
```

The cause of the bug is that the printer doesn't add a space after a node if the previous or
the next node is an inline comment. There's no reason to do that. The only case that warrants
a special behavior is when the current node is an inline comment. Then the printer should add
a hardline after it.