.prettierignore: resolve upwards like .gitignore
Closes https://github.com/prettier/prettier/issues/4081.

When deciding to ignore a file, allow `.prettierignore` files at any directory, not just the current directory.
For example, in this setup:

    .prettignore
    => *.js
    src/
        .prettierignore
        => !regular-module.js
        regular-module.js

`src/regular-module.js` _will_ be formatted, because it was un-ignored by `src/.prettierignore`.

This replaces the default behavior, where previously only the `.prettierignore` in the current directory would take effect.
After this PR, you can run `prettier` from any directory since the resolution is based on the directory of the file.

If passed the `--ignore-path` argument, fall back on the old implementation: only apply rules in the given file, don't resolve upwards.

In other words, you can get the old default behavior by running:

    prettier --ignore-path .prettierignore <arguments>