Gracefully handle nonexistent paths passed to --stdin-filepath.
Previously, if you passed a nonexistent subdirectory to
--stdin-filepath, prettier would simply throw an error instead of using
the path name as information for its config. This occurred because the
editorconfig resolution code used the third party findProjectRoot
function, which throws when given a nonexistent directory.

This adds a pre step that ensures that the path passed to
findProjectRoot is an existing directory.

Fixes #5966.