cli: Handle errors reading stdin
Prettier previously responded to a read error on stdin by sending some
warnings to stderr and exiting successfully with empty output.

$ prettier --stdin-filepath index.js < index.js
(node:23616) UnhandledPromiseRejectionWarning: Error: EACCES: permission denied, read
(node:23616) UnhandledPromiseRejectionWarning: Unhandled promise rejection. This error originated either by throwing inside of an async function without a catch block, or by rejecting a promise which was not handled with .catch(). (rejection id: 1)
(node:23616) [DEP0018] DeprecationWarning: Unhandled promise rejections are deprecated. In the future, promise rejections that are not handled will terminate the Node.js process with a non-zero exit code.
$ echo $?
0

The read error here was caused by a complicated interaction with
snap-confine and AppArmor that has nothing to do with us, but our
response to the error was very bad and could result in data loss when
using prettier-emacs.  Add a handler to report the error correctly:

$ prettier --stdin-filepath index.js < index.js
[error] index.js: Error: EACCES: permission denied, read
$ echo $?
2

Signed-off-by: Anders Kaseorg <andersk@mit.edu>