Suppress “Checking formatting...” message on non-TTY
When Prettier is run by a script that combines the output from dozens
of separate linters, any progress output is just noise that makes it
harder to see whether there are real errors.  Similarly to how we
already suppress printing each filename as it’s checked when
outputting to a non-TTY, also suppress the “Checking formatting...”
and “All matched files use Prettier code style!” messages in this
case.

Signed-off-by: Anders Kaseorg <andersk@mit.edu>