improve legibility of binary expressions with mixed operators
This commit adds parens around mixed operators in binary expressions
when the expression includes mixed division and addition or subtraction
operators. To avoid some issues regarding comments appearing within existing
parens, this will not format binary expressions with leading line comments

e.g.

```js
return (
  // Reason for 42
  42
) / 8 + 2;
```

will not be formatted to avoid a mess with the position of the parens and
comment.