Optimize visitNode by specialising by node type
The new array path avoids creation of a new array iterator of the
entries (which themselves are arrays as well). The object path is
something that has been benchmarked to be a fast way to iterate over
objects.

For an example case I had (about 2000 line .js file) this change brought
preprocess time down from 50ms to about 20ms.