Clarify documentation around proseWrap
These docs were a bit misleading, describing never as "opt-out" when it in fact can result in massive changes in an existing wrapped file is confusing, and referring to the default behavior as "wrapping as-is" is confusing, because Prettier isn't actually doing any wrapping - it's just leaving it as-is. Hopefully this is clearer.