Add assert to utils
By implementing a custom function for assert, we no longer rely on the
Node global polyfill for the assert module. As part of #12144, I
discovered that the node global polyfill is a large contributor to the
file size.

Currently, on main I obtained the following numbers:

```
> yarn build --file=standalone.js --no-minify
> perl -le "map { \$sum += -s } @ARGV; print \$sum" dist/esm/standalone.mjs
< 1207295

> yarn build --file=standalone.js
> perl -le "map { \$sum += -s } @ARGV; print \$sum" dist/esm/standalone.mjs
< 562931
```

With this PR, the numbers change to:

```
> yarn build --file=standalone.js --no-minify
> perl -le "map { \$sum += -s } @ARGV; print \$sum" dist/esm/standalone.mjs
< 1136137

> yarn build --file=standalone.js
> perl -le "map { \$sum += -s } @ARGV; print \$sum" dist/esm/standalone.mjs
< 530682
```

This is a 5.89% and 5.72% size reduction for respectively unminified and
minified standalone bundle output.