Do not use Chalk in error message in main bundle
The CLI uses Chalk for color messages for a better user experience.
However, the standalone bundle does not pull in the CLI and is intended
to be loaded in the browser, which does not have a terminal interface.

By avoiding using Chalk in the main bundle, we obtain the following
numbers:

```
> yarn build --file=standalone.js --no-minify
> perl -le "map { \$sum += -s } @ARGV; print \$sum" dist/esm/standalone.mjs
< 1156128

> yarn build --file=standalone.js
> perl -le "map { \$sum += -s } @ARGV; print \$sum" dist/esm/standalone.mjs
< 539365
```

This is a 4.24% and 4.19% size reduction for respectively unminified and
minified standalone bundle output compared to `main`.