Update precommit Option 1 setup
As for a while now, lint-staged does not recommend anymore using ```npx mrm lint-stage``` because of some issues when installing. I have faced the same problems with it too, and the new way of installing and setting it up worked for me.
It was changed in (this commit)[https://github.com/okonet/lint-staged/blob/5bf1f188f6f6172fc91c79522fa870d887eceba4/README.md#installation-and-setup] and still remains like that.