Fix infinite recursion issue for new Array nodes
When an array is seen it recursively calls a function which checks if the current char position has a "," starting at the nodes end loc position, if it doesn't have a "," it looks at the next token and so on. However in the case when the array is added by a codemod the start position is the start of the file and if its simple enough there may not be a "," in the file so it just keeps recursing until it hits the max call frame stack size and fails. This fix checks if the node has an empty size (which should not normally be possible) and aborts if it is. Additionally this change switches from recursive to iterative and backstops the iteration via the file's text size.