fix: handle long globs
prettier wraps lstat to look for cases where a file does not exist...
When it does not exist, it will try treating the argument as a glob.

For very long globs, the system will return ENAMETOOLONG instead of ENOENT.
This is still a case where the file does not exist and where a glob could
exist.

Instead of letting that error bubble up as an uncaught exception,
swallow it in the same manner as ENOENT.