docs: Use git add -z and xargs -0 in pre-commit script
git-add(1) accepts a `-z` flag that makes it use NUL characters as pathname
separators instead of newlines.
This is suitable with the -0 flag of xargs(1), and escaping spaces with
sed is no longer needed. I also added the `--` flag to both git-add and
prettier, to indicate the end of options and the start of input files.

Signed-off-by: Walid B <76976556+mwb-u@users.noreply.github.com>