Fix indentation for starred block comments in JSX
This patch improves Prettier's handling for multi-line block comments
where each line starts with an asterisk, when those comments appear in a
JSX expression, like this:

```jsx
<div>
  {/*
    * starred
    * comment
    */}
</div>;
```

Prettier already has nice special handling for this in the non-JSX
context: the asterisks are made to line up by adding an extra space at
the start of each non-initial line. In a JSX expression node, though, we
need two spaces, because the prefix is `{/*` and not just `/*`.

This builds on #2470, which left this test case for future work:
https://github.com/prettier/prettier/pull/2470#issuecomment-334964986

Fixes #17487.