fix: avoid closing files multiple times
Fixes prettier/prettier-cli#75.

The readlines library already closes the file descriptor when it reaches
the end of the file, so we were sometimes calling `closeSync` twice on
the same descriptor (via our `finally` block).

Instead, we can let readlines open and close the file, and only close if
it didn't do it itself.